Create a GC Chart Visualization

Generates an interactive GC chart for genomic data.

GC_chart(
  data,
  start = "start",
  stop = "stop",
  cluster = NULL,
  group = NULL,
  width = "100%",
  height = "400px",
  elementId = NULL,
  scale_breaks = TRUE,
  scale_break_threshold = 20,
  scale_break_padding = 1
)

Arguments

data

Data frame containing genomic information.

start

Column name that indicates start positions. Default is "start".

stop

Column name that indicates stop positions. Default is "stop".

cluster

Optional column name used for clustering purposes. Default is NULL.

group

Column name used for gene grouping to influence color aesthetics.

width

Width specification for the chart, such as '100%' or 500. Default is unspecified.

height

Height specification for the chart, such as '400px' or 300. Default is unspecified.

elementId

Optional identifier string for the widget. Default is NULL.

scale_breaks

Logical flag indicating if scale breaks should be employed. Default is TRUE.

scale_break_threshold

Numeric value indicating the threshold percentage of the entire range for determining inter-gene regions suitable for scale breaks. Default is unspecified.

scale_break_padding

Numeric value indicating the padding percentage of the entire range on either side of a scale break. Default is unspecified.

Value

A GC chart widget.

Examples

genes_data <- data.frame(
  start = c(10, 50, 90, 130, 170, 210),
  stop = c(40, 80, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'A', 'B', 'B', 'A', 'C'),
  cluster = c(1, 1, 1, 2, 2, 2)
)
GC_chart(genes_data, group = "group", cluster = "cluster") %>% GC_labels("name")