Update Settings of a GC Chart Item

This function modifies the settings of specified clusters within a GC chart.

GC_item(GC_chart, setting = NULL, cluster = NULL, ...)

Arguments

GC_chart

A GC chart object to be updated.

setting

Character. The setting within the cluster to be modified.

cluster

Numeric or character vector. Cluster(s) within the GC chart to be updated.

...

Additional arguments to modify the settings of the cluster.

Value

An updated GC chart object.

Examples

genes_data <- data.frame(
  start = c(10, 50, 90, 130, 170, 210),
  stop = c(40, 80, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 2', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'A', 'B', 'B', 'A', 'C'),
  cluster = c(1, 1, 1, 2, 2, 2)
)

# Update style of a specific cluster
GC_chart(genes_data, cluster ="cluster", group = "group") %>%
GC_labels("name") %>%
GC_item(
  "labels",
  cluster = 2,
  fontSize = "16px"
)

# Update style of specific items within a specific cluster
GC_chart(genes_data, cluster ="cluster", group = "group") %>%
GC_labels("name") %>%
GC_item(
  "labels",
  cluster = 2,
  itemStyle = list(
    list(index = 0, fontSize = "10px"),  # Index 0 due to JS array
    list(index = 2, fontSize = "16px")
  )
)