Set Legend for a Gene Chart

This function configures the legend for a gene chart. It can toggle the legend's visibility, apply a color scheme, assign custom colors, and set custom labels for the legend entries.

GC_legend(
  GC_chart,
  group = NULL,
  show = TRUE,
  colorScheme = NULL,
  customColors = NULL,
  labels = NULL,
  ...
)

Arguments

GC_chart

The gene chart object to be modified.

group

Optional; character or NULL, the groups to include in the legend. If NULL, the groups are taken from the 'group' attribute of the 'GC_chart' object.

show

Logical, whether to display the legend or not.

colorScheme

Optional; character or NULL, the name of the color scheme to apply to the legend. This can be one of D3.js's built-in color schemes (e.g., "schemeCategory10", "schemeAccent", "schemeTableau10").

customColors

Optional; list or NULL, custom colors to apply to the legend entries.

labels

Optional; character or NULL, custom labels for the legend entries.

...

Additional arguments to be passed to the legend configuration.

Value

Returns the gene chart object with the legend configured.

Examples

genes_data <- data.frame(
  start = c(10, 90, 130, 170, 210),
  stop = c(40, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'B', 'B', 'A', 'C')
)

# Assuming GC_chart is a function that creates a gene chart object
GC_chart(genes_data, cluster = "cluster", group = "group") %>%
  GC_legend(
    position = "top", #bottom
    orientation = "horizontal", #vertical
    x = 10,
    y = 10,
    adjustHeight = TRUE,
    labels = NULL, # c('Group A', 'Group B', 'Group C'),
    legendOptions = list(
      cursor = "pointer",
      colorScheme = NULL,
      customColors = NULL # c("red", "green", "orange")
      # Additional styles
    ),
    legendTextOptions = list(
      cursor = "pointer",
      textAnchor = "start",
      dy = ".35em",
      fontSize = "12px",
      fontFamily = "sans-serif",
      fill = "black"
      # Additional styles
    ),
  )
#> Error in GC_chart(genes_data, cluster = "cluster", group = "group"): cluster column not found in data

# Configure the legend of the gene chart
gene_chart_modified <- gene_chart %>%
  GC_legend(
    show = TRUE,
    colorScheme = 'schemeCategory10', # Example color scheme
    customColors = list('A' = 'blue', 'B' = 'green', 'C' = 'red'),
    labels = c('Group A', 'Group B', 'Group C')
  )
#> Error in GC_legend(., show = TRUE, colorScheme = "schemeCategory10", customColors = list(A = "blue",     B = "green", C = "red"), labels = c("Group A", "Group B",     "Group C")): object 'gene_chart' not found