Set Legend for a Gene Chart

This function configures the legend for a gene chart. It allows toggling the legend's visibility, setting a background color, and assigning custom labels for the legend entries. The function can also handle additional customizations through various options.

Usage

GC_legend(
  GC_chart,
  group = NULL,
  show = TRUE,
  backgroundColor = "#0000",
  labels = NULL,
  legendOptions = list(),
  legendTextOptions = list(),
  ...
)

Arguments

GC_chart

The gene chart object to be modified.

group

Optional; character or NULL, specifies the groups to include in the legend. If NULL, groups are taken from the 'group' attribute of the 'GC_chart' object.

show

Logical, specifies whether to display the legend.

backgroundColor

String, the background color of the legend.

labels

Optional; character vector or NULL, custom labels for the legend entries.

legendOptions

List, additional options for the legend.

legendTextOptions

List, additional text options for the legend.

...

Additional arguments to be passed to the legend configuration.

Value

Returns the modified gene chart object with the legend configured.

Examples

genes_data <- data.frame(
  start = c(10, 90, 130, 170, 210),
  stop = c(40, 120, 160, 200, 240),
  name = c('Gene 1', 'Gene 3', 'Gene 4', 'Gene 5', 'Gene 6'),
  group = c('A', 'B', 'B', 'A', 'C'),
  cluster = c('A', 'A', 'A', 'B', 'B')
)

# Customize legend
GC_chart(genes_data, cluster = "cluster", group = "group") %>%
  GC_legend(
    position = "top", #bottom
    orientation = "horizontal", #vertical
    x = 10,
    y = 10,
    adjustHeight = TRUE,
    backgroundColor = "#0000",
    labels = NULL, # c('Group A', 'Group B', 'Group C'),
    legendOptions = list(
      cursor = "pointer",
      colorScheme = NULL,
      customColors = NULL # c("red", "green", "orange")
      # Additional styles
    ),
    legendTextOptions = list(
      cursor = "pointer",
      textAnchor = "start",
      dy = ".35em",
      fontSize = "12px",
      fontFamily = "sans-serif",
      fill = "black"
      # Additional styles
    )
  )