Update Scale of a GC Chart Cluster

Modify the scale settings for specified clusters within a GC chart.

Usage

GC_scale(
  GC_chart,
  cluster = NULL,
  start = NULL,
  stop = NULL,
  breaks = list(),
  ...
)

Arguments

GC_chart

A GC chart object.

cluster

Numeric or character vector. Clusters in the GC chart to update.

start

Numeric vector. Starting points for the scales.

stop

Numeric vector. Stopping points for the scales.

breaks

List. Settings for the scale breaks.

...

Additional arguments for scale settings.

Value

Updated GC chart with new scale settings.

Examples

genes_data <- data.frame(
start = c(100, 1000, 2000),
stop = c(150, 1500, 2500),
name = c( 'Gene 4', 'Gene 5', 'Gene 6'),
group = c( 'B', 'A', 'C'),
cluster = c(2, 2, 2)
)

# Set manual start and stop values for scales domain
GC_chart(genes_data, cluster ="cluster", group = "group") %>%
GC_scale(
  start = 1,
  stop = 2600,
  breaks =
  list(
    list(start = 160, stop = 900),
    list(start = 1600, stop = 1900)
      )
  )